[ui] Better handling of multiple errors.
We cancel client-side sentences on error, without waiting for the
cancellation commands as to avoid races and error cascades.
This means that a cancel answer may arrive but be redundant as the
span is already gone. We still want to issue the cancel command due to
resilient mode commands.

We also maintain an error list allowing for multiple errors to be
signalled from Coq.

We need more work and a bit of refactoring in this area.