[jscoq] Use JSON serializers from SerAPI for Coq types.
We depend on serlib to serialize Coq data structures and remove the
duplication present in jsCoq. This greatly expands the range of Coq's
types available for sending as JSON.

Next commit will add a query to showcase UI grabbing the full AST for
a sentence, which allows pretty cool things UI-wise.

For now, we include serlib as a module, which is a bit cumbersome,
however once we release serapi 0.7.0 we should be able to just depend
on the coq-serapi OPAM package.

Closes #89