[build] Added JsCoq.is_npm flag.
Needed to account for the location of `node_modules/` in NPM's (current)
directory layout.
When `npm install` is used to install jsCoq, the directory `jscoq` will be
placed alongside its dependencies, rather than as their parent.

The flag is false in the source tree an "regular" dist, and is set to true by
`make dist-npm`.