[addon] Sharing content with Hastebin.
This is phrased as an "addon"; I think that not only libraries can be
"addons" but this kind of interop UI functionality should be as well.

It does not mean that UI addons should be built outside of the jsCoq
source tree like mathcomp & co.
If they are small enough, they are ok where they are in `ui-js/addon`.