[jscoq] Make protocol interpreter independent of worker code.
This is a small refactoring that places the protocol interpreter into
its own file. This is done in preparation for the new printer.

For now I didn't complicate things much, and I have kept the
interpreter access to the `jsoo` libs, tho it would be easy to
actually make it independent by handling the interrupt code in the
worker side [by layering the protocols]

Note that in this setup we are actually architecturally very close
to just replacing our protocol by the SerAPI protocol, while I don't
think we should do that, IMHO it is important to keep the component
structure consistent about projects as SerAPI wants to reuse the
library manager for example in its jsoo version.

Another interesting TODO here would be to actually improve our worker
handling by moving some of our tricks into jsoo upstream.