[jscoq] Preliminary upgrade to Coq 8.16
This is broken due to `load_plugin` not being implemented yet.

We need to think about how to fix that, likely just updating our
library manager to include `META` files for the plugins should work,
but we need to find out.

The main relevant change for us in Coq 8.16 is that Coq now loads
plugins using the `Fl_dynload` API.

So now Coq, will make a request for `coq-core.plugins.ltac` instead of
`ltac_plugin.cma`.

`Fl_dynload` is a wrapper over `Dynlink.loadfile`, but will check
`META` files for dependencies, and if plugins `A` depends on `B`, will
load `B` first.

This may be a problem for us, we need to figure how to handle
that.

One approach couuld be to replace the `Dynlink.dynload` stub with the
proper javascript implementation so the `Fl_dynload` uses it implicitly.

Another (temporal) alternative to get this working is just to ignore
deps and implement `Ml_top.load_plugin` with a mapping. I have done
that as Coq 8.16 still includes a "legacy" mode, which should work for
some plugins, but not for all of them (coq-elpi for example requires this).

The 3rd approach would be for us to reimplement `Fl_dynload` by
locating the META files and parsing them, or storing this metadata in
some other fashion.