[js] [refactor] Remove use of JsCoq global in components.
Instead of using a JsCoq global, we pass the needed parameters to the
constructors so the components types are more functional.

The access to the global was mostly used to determine the backend, but
IMHO this is morally wrong as nothing should preclude a manager
creating N worker objects with different backends.

Moreover, most of the backend case analysis should go away once we
improve backend handling.

This is also a step towards frontend / backend separation.

Next step is indeed to isolate the backend options in their own
interface.