[protocol] Version the document to avoid diagnostics race propagation.
I'm unsure why that happens, but there is a race condition with
transactions coming from diagnostics and transactions coming from
edit.

In particular, a diagnostic transaction could be applied after
an edit transaction, then reverting it. This was noticeable.

We add a version tag to diagnostics and we will refuse to apply
transactions if the current document version is higher.

We also make the input format a parameter of `NewDoc` for now.