[build] Move frontend build to esbuild, always pack.
This is a large build-time improvement, moreover we can get rid of a
lot of troublesome code regarding how to manage assets.

`esbuild` has proven a great tool so far.

Tested with both backends.

We still keep the CLI and the WASM backend on webpack, the idea is
to move them to esbuild in subsequent commits.

CLI should use a native Coq instead, it is super-slow.

We also disabled the `collab` and `ide-project` builds for now,
re-adding before merging hopefully.

The compatibility story with user .html pages needs to be revised,
IMHO as this is a foundation for the LSP backend this is the time to
drop support for some patterns we had in the past.