[jscoq] Unification of package manager among backends.
For a while, we had two package management models, one for `wacoq`,
where the worker loaded a zipped package and registered it, and one
for `jscoq`, where the frontend did that work and the put the files in
the worker.

IIANM, this was introduced when `.coq-pkgs` zipped files were, as we
didn't develop a `.zip` binding for the JSOO worker.

In this PR, we will unify both models by providing such a
backend (using the OCaml `zipc` library).

Thus, workers now always handle `.coq-pkgs` in an uniform way, and so
does the frontend.

This model makes sense because as things stand, it seems more
convenient that the worker "owns" the filesystem, for example to avoid
round-trips when auto-building projects (as the build logic resides in
the server).

But indeed, several designs are possible here, but for example in LSP,
the build system tends to live in the server.

Another goal of this PR is to move loadpath handling to the server,
and hopefully allow more standard loadpath handling in general.

There are many more design questions to handle, this is WIP for now.